Text-Generate-by-Gemini-main/Text-Generate-by-Gemini-main/requirements.txt
fastapi[standard]
uvicorn
python-dotenv
google-generativeai 

